+++ src/qp/utils/matpenalized.c
@@ -134,7 +134,7 @@ static PetscErrorCode MatPenalizedUpdatePenalty_Penalty(Mat Arho, PetscReal rho_
   PetscFunctionBegin;
   PetscCall(MatShellGetContext(Arho, (void *)&ctx));
   rho_new = ctx->rho * rho_update;
-  PetscCall(PetscInfo(permon, "updating rho := %.4e*%.4e = %.4e\n", ctx->rho, rho_update, rho_new));
+  PetscCall(PetscInfo(permon, "updating rho := %.4e*%.4e = %.4e\n", (double)ctx->rho, (double)rho_update, rho_new));
   ctx->rho = rho_new;
   PetscFunctionReturn(PETSC_SUCCESS);
 }
